Refactor audio features and improve documentation
Enhance audio recording, playback, and typing animation functionality in the React application. Key updates include:

- Updated README.md to reflect changes in hooks usage.
- Refactored useAudioPlayback.ts to implement new audio playback logic with error handling.
- Improved useAudioRecording.ts for better state management and added cleanup functionality.
- Enhanced useTypingAnimation.ts for a better user experience.
- Integrated new hooks in InterviewChat.tsx and InterviewControls.tsx for managing audio states and transcription.
- Made minor layout and error handling adjustments in ConductJob.tsx for consistency.
